+++ internal/dbsqlc/query.sql.go
@@ -77,16 +77,8 @@ func (q *Queries) JobCountByQueueAndState(ctx context.Context, db DBTX, queueNam
 }
 
 const jobCountByState = `-- name: JobCountByState :many
-SELECT
-  state,
-  count(*)
-FROM
-  river_job
-WHERE
-  queue IS NOT NULL AND
-  priority > 0 AND
-  scheduled_at IS NOT NULL AND
-  id IS NOT NULL
+SELECT state, count(*)
+FROM river_job
 GROUP BY state
 `